requirements.txt
flask
python-barcode
pillow